+++ src/lib.rs
@@ -1,18 +1,23 @@
 #![doc = include_str!("../README.md")]
 #![warn(missing_docs)]
 
+use std::fmt::Display;
 use std::ops::Range;
 
 mod name_resolver;
-mod rslint;
 mod scope_index;
 mod scope_name;
 mod source;
+mod swc;
 
 pub use name_resolver::NameResolver;
 pub use scope_index::{ScopeIndex, ScopeIndexError, ScopeLookupResult};
 pub use scope_name::{NameComponent, ScopeName};
 pub use source::{SourceContext, SourceContextError, SourcePosition};
+use swc_common::Spanned;
+
+/// The Scopes extracted from a piece of JS Code.
+pub type Scopes = Vec<(Range<u32>, Option<ScopeName>)>;
 
 /// Extracts function scopes from the given JS-like `src`.
 ///
@@ -34,6 +39,7 @@ pub use source::{SourceContext, SourceContextError, SourcePosition};
 /// let src = "const arrowFnExpr = (a) => a; function namedFnDecl() {}";
 /// //                arrowFnExpr -^------^  ^------namedFnDecl------^
 /// let mut scopes: Vec<_> = js_source_scopes::extract_scope_names(src)
+///     .unwrap()
 ///     .into_iter()
 ///     .map(|res| {
 ///         let components = res.1.map(|n| n.components().map(|c| {
@@ -50,31 +56,37 @@ pub use source::{SourceContext, SourceContextError, SourcePosition};
 /// assert_eq!(scopes, expected);
 /// ```
 #[tracing::instrument(level = "trace", skip_all)]
-pub fn extract_scope_names(src: &str) -> Vec<(Range<u32>, Option<ScopeName>)> {
-    rslint::parse_with_rslint(src)
+pub fn extract_scope_names(src: &str) -> Result<Scopes, ParseError> {
+    let mut scopes = swc::parse_with_swc(src).map_err(|e| ParseError { inner: e })?;
+
+    // filter out empty names
+    for scope in &mut scopes {
+        if let Some(ref name) = scope.1 {
+            if name.components.is_empty() {
+                scope.1 = None;
+            }
+        }
+    }
+
+    Ok(scopes)
 }
 
-// TODO: maybe see if swc makes scope extraction easier / faster ?
-/*mod swc {
-    use swc_ecma_parser::lexer::Lexer;
-    use swc_ecma_parser::{Parser, StringInput, TsConfig};
+/// An error parsing the JS Source provided to [`extract_scope_names`].
+#[derive(Debug)]
+pub struct ParseError {
+    inner: swc::ParseError,
+}
 
-    pub fn parse_with_swc(src: &str) {
-        swc_ecma_parser::parse_file_as_module();
-
-        let source = SourceFile;
-
-        let mut parser = Parser::new(
-            swc_ecma_parser::Syntax::Typescript(TsConfig {
-                tsx: true,
-                decorators: true,
-                dts: true,
-                no_early_errors: true,
-            }),
-            StringInput::from(src),
-            None,
-        );
-
-        let module = parser.parse_module().unwrap();
+impl Display for ParseError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let span = self.inner.span();
+        f.write_fmt(format_args!(
+            "{}:{}:{}",
+            span.lo.0,
+            span.hi.0,
+            self.inner.kind().msg()
+        ))
     }
-}*/
+}
+
+impl std::error::Error for ParseError {}